requirements.txt
httpx>=0.27.0
Pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
aiofiles>=23.0.0
